% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{apply_batch_correction}
\alias{apply_batch_correction}
\title{Apply batch correction using limma or linear regression fallback}
\usage{
apply_batch_correction(Mval, cov_mat, use_limma = TRUE)
}
\arguments{
\item{Mval}{features x samples matrix}

\item{cov_mat}{matrix (samples x covariate-columns)}

\item{use_limma}{logical}
}
\value{
corrected matrix (same dimnames)
}
\description{
Apply batch correction using limma or linear regression fallback
}
