% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assocComp_df}
\alias{assocComp_df}
\title{Association between PCs and sample annotations (assocComp-like)}
\usage{
assocComp_df(data_df, sampleAnnotation)
}
\arguments{
\item{data_df}{numeric matrix/data.frame where rows = samples, cols = features}

\item{sampleAnnotation}{data.frame where rows = samples (same order) and columns = annotations}
}
\description{
Association between PCs and sample annotations (assocComp-like)
}
