% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_analysis.R
\name{compute_sample_means}
\alias{compute_sample_means}
\title{Compute per-sample mean methylation (or metric)}
\usage{
compute_sample_means(mat)
}
\arguments{
\item{mat}{numeric matrix of values (rows = features, cols = samples)}
}
\value{
numeric vector of mean per sample
}
\description{
Compute per-sample mean methylation (or metric)
}
