% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_wrappers.R
\name{plot_lda_gg}
\alias{plot_lda_gg}
\title{LDA plotting helper (fixed)}
\usage{
plot_lda_gg(
  lda,
  metadata,
  class_col,
  dim_axes = c(1, 2),
  color_by = NULL,
  shape_by = NULL,
  label_by = NULL,
  label_size = 3,
  title = "LDA plot",
  point_size = 3,
  alpha = 0.8,
  theme = ggplot2::theme_minimal()
)
}
\description{
LDA plotting helper (fixed)
}
