% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_wrappers.R
\name{plot_prcomp_gg}
\alias{plot_prcomp_gg}
\title{PCA plotting helper using ggplot2 (with feature count in title)}
\usage{
plot_prcomp_gg(
  pca,
  metadata = NULL,
  pc_axes = c(1, 2),
  color_by = NULL,
  shape_by = NULL,
  size_by = NULL,
  label_by = NULL,
  label_size = 3,
  repel = TRUE,
  draw_by = NULL,
  ellipse_fill = FALSE,
  ellipse_alpha = 0.15,
  ellipse_size = 1,
  title = "PCA plot",
  point_size = 3,
  alpha = 0.8,
  theme = ggplot2::theme_minimal()
)
}
\description{
PCA plotting helper using ggplot2 (with feature count in title)
}
