% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_analysis.R
\name{prepare_metadata_df}
\alias{prepare_metadata_df}
\title{Prepare metadata dataframe with sample means}
\usage{
prepare_metadata_df(sample_means, metadata)
}
\arguments{
\item{sample_means}{numeric vector of per-sample means}

\item{metadata}{data.frame containing covariates (rows = samples)}
}
\value{
data.frame with mean metric and covariates
}
\description{
Prepare metadata dataframe with sample means
}
