% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_prep.R
\name{prepare_methylation_matrix}
\alias{prepare_methylation_matrix}
\title{Prepare methylation percentage matrix from a methylKit object
Wrapper that calls small subfunctions to remove duplicates, filter samples, and impute}
\usage{
prepare_methylation_matrix(meth_obj, feature_name = "feature")
}
\arguments{
\item{meth_obj}{methylKit object (e.g. methylBase, methylRawList)}

\item{feature_name}{name used in console messages}
}
\description{
Prepare methylation percentage matrix from a methylKit object
Wrapper that calls small subfunctions to remove duplicates, filter samples, and impute
}
