% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter plot with Spearman correlation annotation}
\usage{
scatter_plot(df, x, y, xlab = NULL, ylab = NULL, title = NULL)
}
\arguments{
\item{df}{data.frame}

\item{x}{unquoted column name for x axis}

\item{y}{unquoted column name for y axis}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}
}
\description{
Scatter plot with Spearman correlation annotation
}
