% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{test_assoc_pc_annotation}
\alias{test_assoc_pc_annotation}
\title{Robust test association for a single PC vs annotation}
\usage{
test_assoc_pc_annotation(pc, annotation)
}
\description{
Returns named vector c(pval=..., effect=...)
For numeric: pval from cor.test, effect = correlation estimate
For factor: pval from aov, effect = eta^2 (proportion explained)
}
